MI-39: Added git commit hook
