+++ package.json
@@ -12,7 +12,8 @@
     "format": "prettier . --check",
     "format:write": "prettier . --write",
     "check-types": "nx affected -t check-types",
-    "check-types:all": "nx run-many -t check-types"
+    "check-types:all": "nx run-many -t check-types",
+    "prepare": "[ -d .git ] && git config core.hooksPath '.git-hooks' || true"
   },
   "repository": {
     "type": "git",
@@ -53,6 +54,7 @@
     "packages/*"
   ],
   "nx": {
+    "name": "microservice-development-utilities",
     "includedScripts": [],
     "targets": {
       "start-local-registry": {
