+++ packages/microservice-util-lib/src/get-aws-id-from-arn/get-aws-id-from-arn.ts
