+++ packages/microservice-util-lib/src/has-properties-defined/has-properties-defined.ts
