+++ packages/microservice-util-lib/src/index.ts
@@ -1,10 +1,10 @@
-import chunkBy from './chunkBy/chunkBy';
+import chunkBy from './chunk-by/chunk-by';
 import remap, { Remap, ObjectMap } from './remap/remap';
-import retryWrapper, { RetryConfig } from './retryWrapper/retryWrapper';
-import fetchSsmParams from './fetchSsmParams/fetchSsmParams';
+import retryWrapper, { RetryConfig } from './retry-wrapper/retry-wrapper';
+import fetchSsmParams from './fetch-ssm-params/fetch-ssm-params';
 import S3Dao from './s3/s3';
-import hasDefinedProperties from './hasPropertiesDefined/hasPropertiesDefined';
-import getAwsIdFromArn from './getAwsIdFromArn/getAwsIdFromArn';
+import hasDefinedProperties from './has-properties-defined/has-properties-defined';
+import getAwsIdFromArn from './get-aws-id-from-arn/get-aws-id-from-arn';
 
 export type { Remap, ObjectMap, RetryConfig, S3Dao };
