+++ packages/microservice-util-lib/src/retry-wrapper/retry-wrapper.test.ts
@@ -1,4 +1,4 @@
-import retryWrapper from './retryWrapper';
+import retryWrapper from './retry-wrapper';
 
 describe('retryWrapper', () => {
     it("returns function's result", async () => {
